Create the markup for the user section
Inlcuded all the required images and icons for this section.
Created a new css class for section elements witin the page block that centers the section on the page and also defines a max width.
Applied the new css rule to the user section.
Imported the css file into 'index.css'.